dom/example_unstable.html

<!DOCTYPE html>
<title>Unstable test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(Math.random() > 0.5)
})
</script>
